Graphs Populate Properly
I realized the error in creating "new" graphs was the "update graphs" problem. Instead of creating new, we should update if and only if an update is necessary (such as a new voter login or new vote casted).
This was done by allowing one update per chart per voter/vote; which is why the integration methods were moved into an if statement.
The postgres BarCharts now display/update/switch, but there is an issue regarding the presentation of the axis (a backend/integration issue).

The user is now required to  acknowledge that they are "Logging Out" before returning to the LoginPane. This is currently cosmetic and not functional, in concerns to the backend.

Finally, as agreed by the group, if the member responsible for integration and the showing of the charts, were unable to or did fulfill this responsibility by the time the group thought necessary to do without him, we would remove his credit from the group. As a soft motion toward this, I have added a maybe in front of this member's name on the HomePane.